Fixed an issue with the peers not showing online
This fixes an issue with the peers showing as offline after ~3 minutes
which is the report threshold for being considered online.

The problem was due to the reuse of the WG client, to get the most up to
date stats for a peer you need to create a new client.